+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 ### Added
 - Password support in `shell.copy`
+- `Board.cleanup`, `Channel.register_cleanup`, ability to register a hook
+  for cleaning a channel.  Might help if some lockfiles are kept when TBot
+  just kills a connection.
 
 ### Fixed
 - Stdout showing password prompt late